Pin all metrics crate dependencies
We depend on the alpha version, where breaking changes are occurring
regularly. However, the metrics crate does not pin exact versions of
its dependencies, which can break our CI if a new alpha dependency is
released with breaking changes. Fortunately the metrics crate only has
a single metrics-* dependency, so we can just pin that as well.